+++ CHANGELOG.md
@@ -91,5 +91,6 @@
 * Updated Unity Input Mapper config modal message to reflect the new reality
 * Added AnimateQuadScript and TextureAssignScript (hacky/experimental dynamic texture assignment)
 * Enabled incremental GC
+* Added CollectGarbage API to CoreUtils
 * Implemented dialogue trace (WIP)
 * Implemented lighting for sprite weapon viewmodels (WIP)